chore : restructure directory and refactor files
