+++ deepTextRecognitionBenchmark/create_lmdb_dataset.py
