+++ deepTextRecognitionBenchmark/dataset.py
