+++ deepTextRecognitionBenchmark/demo.py
