+++ deepTextRecognitionBenchmark/modules/feature_extraction.py
