+++ deepTextRecognitionBenchmark/modules/prediction.py
