+++ deepTextRecognitionBenchmark/modules/sequence_modeling.py
