+++ deepTextRecognitionBenchmark/test.py
