+++ deepTextRecognitionBenchmark/train.py
