+++ deepTextRecognitionBenchmark/utils.py
